Añadido el ABM Ingredientes y Etiquetas

+++ MiLibroDeRecetas/Back/Ingrediente.cs
@@ -10,7 +10,7 @@ namespace Back
     {
         public int Id { get; set; }
         public string Nombre { get; set; }
-        public int Calorias { get; set; }
+        public double Calorias { get; set; }
         public string Tipo { get; set; }
     }
 }
